photomanagement/pyproject.toml
[tool.poetry]
name = "photomanagement"
version = "0.1.0"
description = "A package providing functionality for managing your photos."
authors = [
    "Chaniel Abou-Ezzi",
    "Anh Le",
    "Anh Trinh",
    "Kailin Luo",
    "Joseph Schwab"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chromadb = "^0.5.20"
ollama = "^0.3.3"
sentence-transformers = "^3.2.1"
pillow = "^11.0.0"
open-clip-torch = "^2.29.0"
numpy = "^2.1.3"
imagehash = "4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
